+++ src/handlers.rs
@@ -1,0 +1,13 @@
+use axum::{extract::Extension, http::StatusCode, response::IntoResponse, Json};
+use std::sync::Arc;
+
+use crate::repositories::{CreateTodo, TodoRepository};
+
+pub async fn create_todo<T: TodoRepository>(
+    Extension(repository): Extension<Arc<T>>,
+    Json(payload): Json<CreateTodo>,
+) -> impl IntoResponse {
+    let todo = repository.create(payload);
+
+    (StatusCode::CREATED, Json(todo))
+}
