+++ src/main.rs
@@ -1,106 +1,15 @@
-use anyhow::Context;
+mod handlers;
+mod repositories;
+
+use crate::repositories::{TodoRepository, TodoRepositoryForMemory};
 use axum::{
     extract::Extension,
-    http::StatusCode,
-    response::IntoResponse,
     routing::{get, post},
-    Json, Router,
+    Router,
 };
-use serde::{Deserialize, Serialize};
+use handlers::create_todo;
 use std::net::SocketAddr;
-use std::{
-    collections::HashMap,
-    env,
-    sync::{Arc, RwLock},
-};
-use thiserror::Error;
-
-// リポジトリで発生しうるエラーの定義
-#[derive(Debug, Error)]
-enum RepositoryError {
-    #[error("NotFound, id is {0}")]
-    NotFound(i32),
-}
-
-// CRUDの実装をtraitで強制
-// axumでリポジトリを共有するlayer機能を使用するために必要なものを継承
-pub trait TodoRepository: Clone + std::marker::Send + std::marker::Sync + 'static {
-    fn create(&self, payload: CreateTodo) -> Todo;
-    fn find(&self, id: i32) -> Option<Todo>;
-    fn all(&self) -> Vec<Todo>;
-    fn update(&self, id: i32, payload: UpdateTodo) -> anyhow::Result<Todo>;
-    fn delete(&self, id: i32) -> anyhow::Result<()>;
-}
-
-// Todoやそれらの更新に必要なstructを定義
-#[derive(Debug, Deserialize, Serialize, Clone, PartialEq, Eq)]
-pub struct Todo {
-    id: i32,
-    text: String,
-    completed: bool,
-}
-
-#[derive(Debug, Deserialize, Serialize, Clone, PartialEq, Eq)]
-pub struct CreateTodo {
-    text: String,
-}
-
-#[derive(Debug, Deserialize, Serialize, Clone, PartialEq, Eq)]
-pub struct UpdateTodo {
-    text: Option<String>,
-    completed: Option<bool>,
-}
-
-// instance作成メソッドを定義
-impl Todo {
-    fn new(id: i32, text: String, completed: bool) -> Self {
-        Self {
-            id,
-            text,
-            completed: false,
-        }
-    }
-}
-
-type TodoDatas = HashMap<i32, Todo>;
-
-#[derive(Debug, Clone)]
-pub struct TodoRepositoryForMemory {
-    // データアクセスをスレッドセーフにする
-    // RwLock: 可変参照の場合のスレッドアクセスを1つに制限(不偏参照の場合は特に制限なし)
-    store: Arc<RwLock<TodoDatas>>,
-}
-
-impl TodoRepositoryForMemory {
-    pub fn new() -> Self {
-        TodoRepositoryForMemory {
-            store: Arc::default(),
-        }
-    }
-}
-
-// todoRepository traitをTodoRepositoryForMemoryに実装
-impl TodoRepository for TodoRepositoryForMemory {
-    fn create(&self, payload: CreateTodo) -> Todo {
-        todo!();
-    }
-
-    fn find(&self, id: i32) -> Option<Todo> {
-        todo!();
-    }
-
-    fn all(&self) -> Vec<Todo> {
-        todo!();
-    }
-
-    fn update(&self, id: i32, payload: UpdateTodo) -> anyhow::Result<Todo> {
-        todo!();
-    }
-
-    fn delete(&self, id: i32) -> anyhow::Result<()> {
-        todo!();
-    }
-}
+use std::{env, sync::Arc};
 
 #[tokio::main]
 async fn main() {
@@ -135,16 +44,6 @@ fn create_app<T: TodoRepository>(reposiotry: T) -> Router {
         .layer(Extension(Arc::new(reposiotry))) // axumアプリ内でrepositoryを共有できる
 }
 
-// todoを作成
-pub async fn create_todo<T: TodoRepository>(
-    Extension(repository): Extension<Arc<T>>, // 引数の順番になぜか依存がありエラー
-    Json(payload): Json<CreateTodo>,          // Jsonが先に来ているとcompileが通らない
-) -> impl IntoResponse {
-    let todo = repository.create(payload);
-
-    (StatusCode::CREATED, Json(todo))
-}
-
 async fn root() -> &'static str {
     "Hello, world!"
 }
@@ -163,6 +62,7 @@ mod test {
         let repository = TodoRepositoryForMemory::new();
         let req = Request::builder().uri("/").body(Body::empty()).unwrap();
         let res = create_app(repository).oneshot(req).await.unwrap();
+
         let bytes = axum::body::to_bytes(res.into_body(), 128).await.unwrap();
         let body: String = String::from_utf8(bytes.to_vec()).unwrap();
